Applying a better module design pattern.
I've started to apply the Revealing Module pattern to my AHQ models, with the idea that I can better encapsulate the public/private stuff, and implement some form of polymorphism with the DungeonTiles.  Still a work in progress on that front.

Also modified the SVG viewport attribute, so the map will be zoomed in closer (MUCH closer, at this point). Pressing the 'm' key will temporarily zoom out to the full map, so you can see the whole thing. Obviously, pressing keys doesn't work on mobile.